1.add cast op 2.convert cpu quant to tpu 3.alignment bf16 to int8 quant with backend 4.refine int8 preprocess script 5.enable caffe-model tpu quant and int8 preprocess

+++ python/cvi_toolkit/data/preprocess.py
@@ -134,9 +134,8 @@ class preprocess(object):
                 print("not existed {}".format(str(input).rstrip()))
                 return None
 
-            image = image.astype(np.float32)
             image = cv2.resize(image, (self.resize_dims[1], self.resize_dims[0])) # w,h
-
+            image = image.astype(np.float32)
             # opencv read image data format is hwc
             # tranpose here
             image = np.transpose(image, (2, 0, 1))
